Text_Extraction_webapp-streamlit/requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
attrs==20.1.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.14.48
botocore==1.17.48
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipynb==0.5.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.1.3
numpy==1.19.0
opencv-python==4.2.0.34
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
pyarrow==1.0.1
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
streamlit==0.65.2
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.10
validators==0.18.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.0
zipp==3.1.0
streamlit